Tuning
Use colours in all git commands which are capable of coloured output when the
output is to the terminal.

Tell `less` to automatically exit if the entire file can be displayed on the
first screen, display ANSI "colour" escape sequences are output in "raw" form
and Disables sending the termcap initialization and deinitialization strings
to the terminal. Only.