New Postal Address, from today!
Australia Post are closing St Kilda Road Post Shop, so we've moved our
PO Box to be a little closer to home.

Please update your address books accordingly.

Our residential address remains unchanged but please don't send post to
the house as it (the post) will only get wet when it rains (unlike the
UK, post is "delivered" to a hole in the front fence and not through a
slot in the front door).