I am taking part in the Great Cycle Challenge this October to fight kids' cancer!
I'm uncertain of the wisdom of setting myself a 500km cycling target for October when there will be a teeny tiny baby in the house; particularly given I've already entered a 10km run at the Melbourne Marathon on 12 October (I know, sometimes I surprise even myself).

But there we go... we can't spend our lives sitting still, can we?

It's for a great cause, so if you fancy pedalling along with me for a few kilometres in October please give me a holler! Alternatively if you'd rather donate to the Children's Medical Research Institute please go to http://goo.gl/ZOBy2k where you'll find details of how to contribute, along with details of my progress.